+++ src/supplier/supplier.controller.ts
@@ -8,8 +8,8 @@ const em = orm.em;
 
 async function findAll(req: Request, res: Response){
   try{
-    const cities = await em.find(Supplier, {});
-    res.status(200).json({message:'found all cities',data: cities});
+    const products = await em.find(Supplier, {});
+    res.status(200).json({message:'found all products',data: products});
   } catch (error: any) {
     res.status(500).json({message: error.message});
   }
@@ -49,12 +49,23 @@ async function add(req: Request, res: Response) {
   async function update(req: Request, res: Response){
     try{
       const id = req.params.id;
-      const supplier = em.getReference(Supplier, id);
-      em.assign(supplier, req.body);
+      const existingSupplier = await em.findOne(Supplier, { id });
+      if (!existingSupplier) {
+        return res.status(404).json({ message: 'Supplier not found' });
+      }
+  
+      const newCuit = req.body.Cuit;
+      if (newCuit !== existingSupplier.cuit) {
+        const duplicateSupplier = await em.findOne(Supplier, { cuit: newCuit });
+        if (duplicateSupplier) {
+          return res.status(400).json({ message: 'Error', error: 'The new Cuit is already used' });
+        }
+      }
+      em.assign(existingSupplier, req.body);
       await em.flush();
       res
         .status(200)
-        .json({message: 'supplier updated', data: supplier});
+        .json({message: 'supplier updated', data: existingSupplier});
     }
     catch (error: any) {
       res.status(500).json({message: error.message});
@@ -64,7 +75,15 @@ async function add(req: Request, res: Response) {
  async function remove(req: Request, res: Response){
   try{
     const id = req.params.id;
-    const supplier = em.getReference(Supplier, id);
+    const supplier = await em.findOne(Supplier, { id });
+    if (!supplier) {
+      return res.status(404).json({ message: 'Supplier not found' });
+    }
+    const products = await em.find(Product, { supplier });
+    if (products.length > 0) {
+      return res.status(400).json({ message: 'Error', error: 'The supplier has associated products.' });
+    }
+
     await em.removeAndFlush(supplier);
     res
       .status(200)
